Added: 1.Sign out function 2.refactored code
