+++ app/controllers/application.rb
@@ -1,0 +1,4 @@
+get '/' do
+  @peeps = Peep.all
+  erb :index
+end
