+++ app/views/index.erb
@@ -1,3 +1,5 @@
+<br><h1>Hey, What's Peepin'?</h1><br>
+
 <ul>
   <% @peeps.each do |peep| %>
     <li>
@@ -7,6 +9,8 @@
     </li>
   <% end %>
 </ul>
+
+<br>
 
 <peep id="new-peep">
   <form action="/peeps" method="post">
