+++ app/views/welcome.erb
@@ -1,0 +1,10 @@
+<% if current_maker %>
+
+  Welcome, <%= current_maker.name %>.
+
+  <form method="post" action="/sessions">
+    <input type="hidden" name="_method" value="delete">
+    <input type="submit" value="Sign out">
+  </form>
+
+<% end %>
