+++ spec/features/maker_man_spec.rb
@@ -1,11 +1,13 @@
 require 'spec_helper'
+require_relative 'helpers/session'
+
+include SessionHelpers
 
 feature "Maker signs up" do
 
   scenario "as a new user visiting the site" do
     expect{ sign_up }.to change(Maker, :count).by(1)
     expect(page).to have_content("Welcome, Snow White")
-    expect(Maker.first.name).to eq("Snow White")
   end
 
   scenario "with unmatching passwords" do
@@ -24,20 +26,6 @@ feature "Maker signs up" do
     expect(page).to have_content("Email is already taken" || "Username is already taken")
   end
 
-  def sign_up(name = "Snow White",
-              username = "red_apple",
-              email = "snow_white@example.com",
-              password = "seven_dwarfs",
-              password_confirmation = "seven_dwarfs")
-    visit '/makers/new'
-    fill_in :name, :with => name
-    fill_in :username, :with => username
-    fill_in :email, :with => email
-    fill_in :password, :with => password
-    fill_in :password_confirmation, :with => password_confirmation
-    click_button "Sign up"
-  end
-
 end
 
 
@@ -46,6 +34,7 @@ feature "Maker signs in" do
   before(:each) do
     Maker.create( :name => "Snow White",
                   :username => "red_apple",
+                  :email => "snow_white@example.com",
                   :password => "seven_dwarfs",
                   :password_confirmation => "seven_dwarfs")
   end
@@ -64,18 +53,23 @@ feature "Maker signs in" do
     expect(page).not_to have_content("Welcome, Snow White")
   end
 
-def sign_in(username, password)
-  visit '/sessions/new'
-  fill_in 'username', :with => username
-  fill_in 'password', :with => password
-  click_button 'Sign in'
 end
 
+feature "Maker signs out" do
+
+  before(:each) do
+    Maker.create( :name => "Snow White",
+                  :username => "red_apple",
+                  :email => "snow_white@example.com",
+                  :password => "seven_dwarfs",
+                  :password_confirmation => "seven_dwarfs")
+  end
+
+  scenario "while being signed in" do
+    sign_in('red_apple', 'seven_dwarfs')
+    click_button "Sign out"
+    expect(page).to have_content("Good bye!")
+    expect(page).not_to have_content("Welcome, Snow White")
+  end
+
 end
-
-# feature "Maker logs out" do
-
-#   scenario "while being signed in" do
-#   end
-
-# end
